% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avgLoading.R
\name{avgLoading}
\alias{avgLoading}
\title{Calculate average loadings of each cluster}
\usage{
avgLoading(df, k)
}
\arguments{
\item{df}{A data frame. Each row represents principle components from different
training datasets. Each column represents genes used for PCA analysis.}

\item{k}{The number of clusters used for k-means clustering}
}
\value{
A data frame of average loadings. Each column represents cluster
and rows represent genes used for PCA.
}
\description{
Name of the input vector is principle component (pc) from different datasets (ds)
and the integer is the cluster number assigned to each ds.pc.
}
