% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildAvgLoading.R
\name{buildAvgLoading}
\alias{buildAvgLoading}
\title{Calculate average loadings of each cluster}
\usage{
buildAvgLoading(df, k, seed = NULL)
}
\arguments{
\item{df}{A data frame. Each row represents principle components from different
training datasets. Each column represents genes used for PCA analysis.}

\item{k}{The number of clusters used for k-means clustering}

\item{seed}{a random seed}
}
\value{
A list of kmeans clustering results and a data frame of average loadings.
Each column represents cluster and rows represent genes used for PCA.
}
\description{
Name of the input vector is principle component (pc) from different datasets (ds)
and the integer is the cluster number assigned to each ds.pc. This function includes
kmenas clustering, so `set.seed()` before using this function for reproducible
results.
}
