% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{calcSilWidth}
\alias{calcSilWidth}
\title{Calculate silhouette width}
\usage{
calcSilWidth(distance, lvls)
}
\arguments{
\item{distance}{a matrix like object of distance}

\item{lvls}{levels}
}
\value{
data frame with claculated silhouette width
}
\description{
Calculate silhouette width
}
