% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calPC.R
\name{calPC}
\alias{calPC}
\title{PCA on gene expression profile}
\usage{
calPC(x)
}
\arguments{
\item{x}{Numeric data matrix or numeric data frame: each row corresponds to genes, and each column corresponds to samples of the study.}
}
\value{
results of principle components analysis (PCA) in an object of class \code{\link[stats]{prcomp}}
}
\description{
PCA on gene expression profile
}
