% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calScore.R
\name{calScore}
\alias{calScore}
\title{Calculate the score for a new dataset}
\usage{
calScore(dataset, avg.loadings)
}
\arguments{
\item{dataset}{A list of SummarizedExperiment (or ExpressionSet) objects.
Rownames are in 'symbol' format.}

\item{avg.loadings}{Output from `avgLoading` function - a data frame of avaerage
loadings. Each column represents cluster and rows represent genes used for PCA.}
}
\value{
A list containing the score matrices for input datasets. Scores are
assigned to each sample (row) on each cluster (column).
}
\description{
Rownames of dataset and avg.loadings should be gene symbols.
}
