% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commonGene.R
\name{commonGene}
\alias{commonGene}
\title{Common genes from datasets}
\usage{
commonGene(setNames)
}
\arguments{
\item{setNames}{a character vector. Name of the datasets.}
}
\value{
a character vector with the common genes of the datasets.
}
\description{
Common genes from datasets
}
