% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRC_Jaccard_Coef.R
\name{CRC_Jaccard_Coef}
\alias{CRC_Jaccard_Coef}
\title{CRC Jaccard's Coefficient Table}
\usage{
CRC_Jaccard_Coef(res, ds, pc)
}
\arguments{
\item{res}{Output from `kmeans()` clustering}

\item{ds}{the number of datasets used for clutering}

\item{pc}{the number of PCs kept from each dataset}
}
\value{
A data frame of Jaccards' coefficients between your dataset and CRC
data. Each column represents cluster (crc_cl1 ~ crc_cl4, which represent PCSS1
~ PCSS4) and the rows represent PCs from different datasets used in clustering.
}
\description{
Create Jaccard's Coefficient table represents the binary principle components
membership of 8 CRC datasets used in the previous work by Ma *et al.*(2018)
(Genome Biol. 2018 Sep 25;19(1):142. doi: 10.1186/s13059-018-1511-4).
}
