% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{expr.all.mapped}
\alias{expr.all.mapped}
\title{Eight CRC expression datasets}
\format{a list object with 8 matrices.}
\source{
\url{https://bitbucket.org/biobakery/crc-subtyping-paper/src/master/data/eSets/}{curatedCRCData}
}
\usage{
expr.all.mapped
}
\description{
List of length 8, where expressions are centered and scaled on a per-gene level
first, per usual PCA standard. "MEOX2" gene is removed.
}
\references{
Ma et al. (2018) Genome Biol.
(\url{https://www.ncbi.nlm.nih.gov/pubmed/?term=Continuity+of+transcriptomes+among+colorectal+cancer+subtypes+based+on+meta-analysis}{PubMed})
}
\keyword{datasets}
