% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{fillInZero}
\alias{fillInZero}
\title{Filling in zero values for classification}
\usage{
fillInZero(exprs, names)
}
\arguments{
\item{exprs}{an ExpressionSet object}

\item{names}{gene names}
}
\description{
Filling in zero values for classification
}
