% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findStudiesInCluster.R
\name{findStudiesInCluster}
\alias{findStudiesInCluster}
\title{Find the studies contributing each cluster}
\usage{
findStudiesInCluster(PCAmodel, ind = NULL)
}
\arguments{
\item{PCAmodel}{GenomicSignatures object contianing a PCA model}

\item{ind}{a numeric vector containing the index of PC clusters you want to
find related studies. Default is `NULL`.}
}
\value{
If `ind` is not specified, this function will create a binary membership
matrix with the PCs in the raw and the clusters in the column. If `ind` is specificed,
the ouput will be a list of studies in the specified cluster.
}
\description{
Find the studies contributing each cluster
}
