% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmapTable.R
\name{heatmapTable}
\alias{heatmapTable}
\title{Color the matrix in a heatmap-stype}
\usage{
heatmapTable(dat, title, color)
}
\arguments{
\item{dat}{a matrix subjected to add background colors in each slot correlated
to its value - similar to heatmap}

\item{title}{a character string containing the title of the table}

\item{color}{a numeric vector of length 3. Number represents the values assigned
to "grey", "white", and "red", respectively. Default is `c(0, 0.4, 0.8)`.}
}
\value{
a heatmap version of the input matrix
}
\description{
Color the matrix in a heatmap-stype
}
