% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jaccardIndex.R
\name{jaccardIndex}
\alias{jaccardIndex}
\title{Create Jaccard Similarity Table}
\usage{
jaccardIndex(x, y)
}
\arguments{
\item{x}{the Jaccard coefficient table of reference clustering}

\item{y}{the Jaccard coefficient table of test clustering. nrow(y) should be
same as nrow(x)}
}
\value{
A data frame of Jaccard similarity score. Each column represents clusters
subjected to test and the rows represent clusters from the reference clustering.
}
\description{
Calculate Jaccard Similarity Score between the reference cluster and the testing
cluster.
}
