% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTableWithTopPCs.R
\name{makeTableWithTopPCs}
\alias{makeTableWithTopPCs}
\title{Create a table of PCs of datasets}
\usage{
makeTableWithTopPCs(datasets, pc.num = 20, commonGene = NULL)
}
\arguments{
\item{datasets}{a list of data frames. Each data frame represent genes x samples of a study.}

\item{pc.num}{the number of top PCs to keep for each dataset after PCA. Default is 20.}

\item{comonGene}{a character vector of the common genes. Default is 'NULL',
in which case, the common genes of the provided datasets will be used.}
}
\value{
A data frame of principle components (PCs) from each dataset. Rows are
common genes from all datasets and columns represent PCs (defined by \code{pc.num})
from each dataset.
}
\description{
Principle component analysis is performed on each dataset and the table of top
principle components (number defined by `pc.num` argument) are saved.
}
