% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numpc.R
\name{numpc}
\alias{numpc}
\title{Calculate the number of PC to return}
\usage{
numpc(data, method = "elbow", B = 20, seed = NULL)
}
\arguments{
\item{data}{the same data as to be used for PLIER (z-score recommended) or alternatively the result of an svd calculation}

\item{method}{Either "eblow" (fast) or "permutation" (slower, but less heuristic)}

\item{B}{number of permutations}

\item{seed}{a random seed for reproducibility}
}
\value{
an integer. The number of cluster for kmeans clustering calculated using elbow method.
}
\description{
This function estimates the number of 'significant' principle components for
the SVD decomposition. Output from this function is the minimum `k` for `PLIER::PLIER()`.
}
\note{
This function is adopted from `PLIER::num.pc()`.
}
