% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcInCluster.R
\name{pcInCluster}
\alias{pcInCluster}
\title{Create Jaccard's Coefficient Table}
\usage{
pcInCluster(x, k)
}
\arguments{
\item{x}{a data frame with the common genes (row) and the principle components
(PCs) from different datasets (col). The number of column is `the number of PCs
x the number of datasets`}

\item{k}{the number of clusters (k) for `Kmeans` clustering}
}
\value{
A data frame of Jaccards' Coefficient. Each column represents cluster
and rows represent PCs from different datasets.
}
\description{
Create Jaccard's Coefficient Table
}
