% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PCs_10ovc_training}
\alias{PCs_10ovc_training}
\title{Colon cancer (CRC) datasets from curatedCRCData for training}
\format{A list object with 8 character strings}
\source{
\url{https://bitbucket.org/biobakery/crc-subtyping-paper/src/master/data/eSets/}{curatedCRCData}
}
\usage{
PCs_10ovc_training
}
\value{

}
\description{
Colon cancer (CRC) datasets from curatedCRCData for training
}
\references{
Ma et al. (2018) Genome Biol.
(\url{https://www.ncbi.nlm.nih.gov/pubmed/?term=Continuity+of+transcriptomes+among+colorectal+cancer+subtypes+based+on+meta-analysis}{PubMed})
}
\keyword{datasets}
