% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcInCluster.R
\name{PCsInCluster}
\alias{PCsInCluster}
\title{Create Jaccard's Coefficient Table}
\usage{
PCsInCluster(x, k, seed = NULL)
}
\arguments{
\item{x}{a data frame with the common genes (row) and the principle components
(PCs) from different datasets (col). The number of column is `the number of PCs
x the number of datasets`}

\item{k}{the number of clusters for `kmeans` clustering if you provide raw data}

\item{seed}{a random seed}
}
\value{
A data frame of Jaccards' Coefficient. Each column represents cluster
and rows represent PCs from different datasets.
}
\description{
Create Jaccard's Coefficient Table
}
