% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcTable.R
\name{pcTable}
\alias{pcTable}
\title{Create a table of PCs of datasets}
\usage{
pcTable(setNames, pc.num = 20, commonGene = NULL)
}
\arguments{
\item{setNames}{a character vector of name of the datasets.}

\item{pc.num}{the number of top PCs to keep for each dataset after PCA}
}
\value{
a data frame with the common genes among datasets (row) and the number
of PCs defined by pc.num (default = 20) from each dataset (column)
}
\description{
Any gene with NA or -Inf/Inf values is removed, followed by centering. Each
datasets are subset with their commone genes.
Principle component analysis is performed on each dataset assuming each samples as
}
