% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R, R/rmNaInf.R
\name{rmNaInf}
\alias{rmNaInf}
\title{Misc Functions for GenomicSuperSignature package}
\usage{
rmNaInf(x)

rmNaInf(x)
}
\arguments{
\item{x}{a matrix with the expression values}
}
\value{
the updated input matrix where NA and Inf values are removed
}
\description{
Remove missing / Inf values in an expression
}
