% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleScoreHeatmap.R
\name{sampleScoreHeatmap}
\alias{sampleScoreHeatmap}
\title{Plot heatmap of the sample scores}
\usage{
sampleScoreHeatmap(
  score,
  dataName,
  modelName,
  cluster_rows = TRUE,
  cluster_columns = TRUE,
  show_row_names = TRUE,
  show_column_names = TRUE,
  row_names_gp = 0.7,
  column_names_gp = 5,
  ...
)
}
\arguments{
\item{score}{a matrix with samples (row) and PrcompClusters (column). If it
is a simple vector, it will be converted to a one-column matrix.}

\item{dataName}{a character string containing the name of the dataset}

\item{modelName}{a character string containing the average loading used for scoring}

\item{cluster_rows}{a logical value, that controls whether to make cluster on rows. Default is TRUE.}

\item{cluster_columns}{a logical value, that controls whether to make cluster on columns. Default is TRUE.}
}
\value{
a heatmap version of the sample score
}
\description{
Plot heatmap of the sample scores
}
