% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetVariantGenes.R
\name{subsetVariantGenes}
\alias{subsetVariantGenes}
\title{Subset highly variant genes}
\usage{
subsetVariantGenes(x, percent.var)
}
\arguments{
\item{x}{a data frame. Row represents different genes.}

\item{percent.var}{a percentage of top variant genes to subset. Only the multiple
of 10 is allowed (maximum = 100).}
}
\value{
a subset of input data frame, which contains only the top `percent.var`
percentage of genes.
}
\description{
Subset highly variant genes
}
