% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{trainingSets_10ovc}
\alias{trainingSets_10ovc}
\title{Names of ten OVC training datasets}
\format{a list object with 10 character strings}
\source{
\url{https://github.com/waldronlab/curatedOvarianData/tree/master/data}{curatedOvarianData}
}
\usage{
trainingSets_10ovc
}
\description{
A list of the dataset names from curatedOvarianData package, used as a training
dataset to build a model. Eight datasets were processed by Affymetrix, and two
datasets were processed by Agilent (GSE17260_eset, GSE32062.GPL6480_eset)
}
\keyword{datasets}
