% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{trainingSets_8crc}
\alias{trainingSets_8crc}
\title{Names of eight CRC training datasets}
\format{a list object with 8 character strings}
\source{
\url{https://bitbucket.org/biobakery/crc-subtyping-paper/src/master/data/eSets/}{curatedCRCData}
}
\usage{
trainingSets_8crc
}
\value{

}
\description{
Names of eight CRC training datasets
}
\references{
Ma et al. (2018) Genome Biol.
(\url{https://www.ncbi.nlm.nih.gov/pubmed/?term=Continuity+of+transcriptomes+among+colorectal+cancer+subtypes+based+on+meta-analysis}{PubMed})
}
\keyword{datasets}
