% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/val.R
\name{val}
\alias{val}
\title{Validatin of a new dataset}
\usage{
val(dataset, avg.loadings)
}
\arguments{
\item{dataset}{A list of SummarizedExperiment (or ExpressionSet) objects.
Rownames are in 'symbol' format.}

\item{avg.loadings}{output from `avgLoading` function - a data frame of avaerage
loadings. Each column represents cluster and rows represent genes used for PCA.}
}
\value{
a matrix containing the correlation coefficiency between the loadings
of new dataset(s) and pre-calculated average loadings of training datasets. Each
row represents a new dataset for test, and each column represents clusters from
training datasets
}
\description{
Pearson correlation based validation of new datasets against average loadings.
}
