% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validte.R
\name{validate}
\alias{validate}
\title{Validatin of a new dataset}
\usage{
validate(dataset, avg.loadings, level = c("dataset"))
}
\arguments{
\item{dataset}{A list of SummarizedExperiment (or ExpressionSet) objects.
Rownames are in 'symbol' format.}

\item{avg.loadings}{output from `avgLoading` function - a data frame of avaerage
loadings. Each column represents cluster and rows represent genes used for PCA.}
}
\value{
A matrix containing the maximum pearson correlation coefficiency between
the top 8 PCs of the dataset(s) and pre-calculated average loadings of training
datasets. Each row represents a new dataset for test, and each column represents
clusters from training datasets.
}
\description{
Validatin of a new dataset
}
