% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calPC.R
\name{calPC}
\alias{calPC}
\title{Calculate Principle Components}
\usage{
calPC(x)
}
\arguments{
\item{x}{expression count matrix where rows represent genes and columns represent each sample of the study}
}
\value{
results of principle components analysis in an object of class \code{\link[stats]{prcomp}}
}
\description{
Calculate Principle Components
}
