% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calScore.R
\name{calScore}
\alias{calScore}
\title{Calculate the score for a new dataset}
\usage{
calScore(dataset, avg.loadings)
}
\arguments{
\item{dataset}{a list of SummarizedExperiment objects. Rownames are in 'symbol' format.}

\item{avg.loadings}{output from `avgLoading` function - a data frame of avaerage
loadings. Each column represents cluster and rows represent genes used for PCA.}
}
\value{
a matrix containing the correlation coefficiency between the loadings
of new dataset(s) and pre-calculated average loadings of training datasets. Each
row represents a new dataset for test, and each column represents clusters from
training datasets
}
\description{
Calculate the score for a new dataset
}
