% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRC_Jaccard_Coef.R
\name{CRC_Jaccard_Coef}
\alias{CRC_Jaccard_Coef}
\title{CRC Jaccard's Coefficient Table}
\usage{
CRC_Jaccard_Coef(res)
}
\arguments{
\item{res}{Output from `kmeans()` clustering}
}
\value{
A data frame of Jaccards' Coefficient with CRC data. Each column represents
 cluster (crc_cl1-crc_cl4, which represent PCSS1-PCSS4) and the rows represent PCs
 from different datasets used for clustering.
}
\description{
Create Jaccard's Coefficient table using the list of loadings used to create
PCSS1-4 in the previous work by Ma *et al.*(2018)
(Genome Biol. 2018 Sep 25;19(1):142. doi: 10.1186/s13059-018-1511-4)
}
