% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmapTable.R
\name{heatmapTable}
\alias{heatmapTable}
\title{Color the matrix in a heatmap-stype}
\usage{
heatmapTable(dat, row_title, column_title, breaks, colors)
}
\arguments{
\item{dat}{a matrix subjected to add background colors in each slot correlated
to its value - similar to heatmap}

\item{row_title}{a character string containing the title of the row}

\item{column_title}{a character string containing the title of the column}

\item{breaks}{a numeric vector of length 3. Number represents the values assigned
to three colors. Default is `c(0, 0.5, 1)`.}

\item{colors}{a character vector of length 3. Each represents the color assigned
to three breaks. Default is `c("grey", "white", "red")`.}
}
\value{
a heatmap version of the input matrix
}
\description{
Color the matrix in a heatmap-stype
}
