% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmTable.R
\name{hmTable}
\alias{hmTable}
\title{Score table with color}
\usage{
hmTable(dat, title, color)
}
\arguments{
\item{dat}{a matrix containing the correlation coefficiency between the loadings
of new dataset(s) and pre-calculated average loadings of training datasets.}

\item{title}{a character string containing the title of the table}

\item{color}{a numeric vector of length 3.}
}
\value{
a heatmap version of the input matrix
}
\description{
Score table with color
}
