% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcInCluster.R
\name{pcInCluster}
\alias{pcInCluster}
\title{Create Jaccard's Coefficient Table}
\usage{
pcInCluster(res)
}
\arguments{
\item{res}{output from `kmeans()` clustering}
}
\value{
A data frame of Jaccards' Coefficient. Each column represents cluster
and rows represent PCs from different datasets.
}
\description{
Create Jaccard's Coefficient Table
}
