% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcTable.R
\name{pcTable}
\alias{pcTable}
\title{Create a table of PCs of datasets}
\usage{
pcTable(setNames, pc.num = 20, commonGene = NULL)
}
\arguments{
\item{setNames}{a character vector of name of the datasets.}

\item{pc.num}{the number of top PCs to keep for each dataset after PCA. Default is 20.}

\item{comonGene}{a character vector of the common genes. Default is 'NULL',
in which case, the common genes of the provided datasets will be used.}
}
\value{
A data frame of principle components (PCs). Rows are common genes of
datasets and columns represent PCs (defined by \code{pc.num}) from each dataset.
}
\description{
Any gene with NA or -Inf/Inf values is removed, followed by centering. Each
datasets are subset with their commone genes.
Principle component analysis is performed on each dataset assuming each samples as
}
