% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleScoreHeatmap.R
\name{sampleScoreHeatmap}
\alias{sampleScoreHeatmap}
\title{Plot heatmap of the sample scores}
\usage{
sampleScoreHeatmap(dat, dataName, avgLoading, cluster_rows = TRUE,
  cluster_columns = TRUE)
}
\arguments{
\item{dat}{a matrix with samples (row) and PrcompClusters (column)}

\item{dataName}{a character string containing the name of the dataset}

\item{avgLoading}{a character string containing the average loading used for scoring}

\item{cluster_rows}{a logical value, that controls whether to make cluster on rows. Default is TRUE.}

\item{cluster_columns}{a logical value, that controls whether to make cluster on columns. Default is TRUE.}
}
\value{
a heatmap version of the sample score
}
\description{
Plot heatmap of the sample scores
}
