% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validte.R
\name{validate}
\alias{validate}
\title{Validatin of a new dataset}
\usage{
validate(dataset, avgLoading, level = "max")
}
\arguments{
\item{dataset}{Single or a list of SummarizedExperiment (or ExpressionSet) object(s).
Rownames are in 'symbol' format.}

\item{avgLoading}{output from \code{avgLoading} function - a data frame of avaerage
loadings. Each column represents cluster and rows represent genes used for PCA.}

\item{level}{Defibe how to ouput validation in two different forms, \code{c("max", "all")}.
Default is "max", which outputs the matrix containing only the maximum coefficient.
To get the coefficient of all 8 PCs, set this argument as "all".}
}
\value{
A matrix containing the maximum pearson correlation coefficient between
the top 8 PCs of the dataset(s) and pre-calculated average loadings of training
datasets. Each row represents a new dataset for test, and each column represents
clusters from training datasets. If \code{level = "all"}, a list containing the matrices
of the pearson correlation coefficient between all top 8 PCs of the dataset(s) and
avgLoading.
}
\description{
Validatin of a new dataset
}
